requirements.txt
asgiref==3.8.1
bleach==6.2.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
crispy-bootstrap5==2025.4
decouple==0.0.7
Django==5.2.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.2.1
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
